Fix modification detection for SP 2010
We used to infinite loop when detecting modifications with SP 2010. The
fix for that effectively killed modification detection outright. The
real issue is that SP 2010's timeout value is different than 2007 and we
were using too low of a value for SP 2010 to return any results. Thus,
that caused us to page through an infinite number of pages of size zero.