Support non-Public URL Alternate Access Mapping
If the user configures sharepoint.server to be the non-Public URL for a
Zone, then the adaptor will end up contacting a different host for all
webs. Thus, with this change we authenticate with other hosts provided
by the configured SharePoint instance.

This likely is all that is necessary to support Host-named Site
Collections as well, but they are untested.