Reduce calls necessary for attachments
We no longer verify that an attachment exists as a child of the itemId.
If the document exists, it is required to be a child of the itemId per
SharePoint.

The TODO to respondNotFound() for invalid itemIds is not new, since
getContentListItemAttachments() threw an exception if the item did not
exist.