Remove redundant calls to obtain items from cache
This makes it more obvious which cached values are re-used which is
helpful for optimizing. For instance, previously we had considered
swapping the order of the anonymous access calls (so that virtual server
is first, because that is more likely to be in our cache), however, now
it is obvious that doing the swapping yields no benefit or even a loss
because the CachedWeb will be loaded anyway.