Combine SOAP Factory objects
We have two factory objects passed to the adaptor at present, but we are
about to add a third. The method we had worked okay for few numbers, but
as we use more SOAP services it won't scale well because of all the
updating of tests. Therefore, this change combines the factories into a
single interface with only one object passed to the constructor.