Fix spaces in Site names in more places
All the methods of SoapFactory could end up being passed URLs with
spaces, so we need to deal with spaces on each. However, using
spUrlToUri() within SoapFactory is misplaced because that should not be
the SoapFactory's responsibility. Thus, we move spUrlToUri() to all the
callers of the SoapFactory.

Unfortunately that means that testing spaces in Site names becomes much
harder and we remove a test case because it is too hard to replicate at
this time.