Prevent hard-coded words from being indexed
For any words we add to generated HTML pages, we now tell the GSA to not
index them. They can still be used in snippets, but they won't
contribute to scoring.

In addition, we only use user-provided wording for HTML titles since I'm
not certain googleoff functions properly in the title (it may prevent
indexing, but it may still influence scoring) and because those titles
are always shown to the user in full (and so our titles now look better
in search results).