Convert SP-specific mime types to GSA-expected type
SharePoint uses some mime types that are not used anywhere else. The GSA
does not support these mime types so we use the IANA-standard types
instead. Using a mime-type as known by the GSA prevents the GSA from
treating the document as a zip and extracting its contents.

Also, there are some small gaps in the GSA's mime type support, so we do
a similar conversion for those.

The list of initial mime types was based on
http://support.microsoft.com/kb/2661910