Filter Lists with empty default view URLs
Such lists do not cause much trouble now, other than the fact they don't
work and produce an odd link.  Most of what this change is doing is
saying, "We don't support this." In the future, we may add support for
such things (maybe while trying to solve various issues with Views in
general).